+++ crates/cli-tools/src/action.rs
@@ -496,7 +496,7 @@ pub struct RustAppletBuild {
     pub crate_dir: PathBuf,
 
     /// Copies the final artifacts to this directory.
-    #[arg(long, value_name = "DIRECTORY", default_value = "target/wasefire")]
+    #[arg(long, value_name = "DIRECTORY", default_value = "wasefire")]
     #[arg(value_hint = ValueHint::DirPath)]
     pub output_dir: PathBuf,
 
@@ -567,6 +567,7 @@ impl RustAppletBuild {
             None => (format!("wasm32-unknown-unknown/{profile}/{name}.wasm"), "applet.wasm"),
             Some(target) => (format!("{target}/{profile}/lib{name}.a"), "libapplet.a"),
         };
+        fs::create_dir_all(&self.output_dir).await?;
         let applet = self.output_dir.join(dst);
         if fs::copy_if_changed(target_dir.join(src), &applet).await? && dst.ends_with(".wasm") {
             optimize_wasm(&applet, self.opt_level).await?;
@@ -596,6 +597,7 @@ impl RustAppletTest {
         let mut cargo = Command::new("cargo");
         cargo.args(["test", "--features=test"]);
         cargo.args(&self.cargo);
+        cargo.current_dir(&self.crate_dir);
         cmd::replace(cargo)
     }
 }